Removed some warnings, added some comments
git-svn-id: http://svn.esdi-humboldt.eu/repo/humboldt2/trunk@3689 5d385326-3dfc-11dd-8f4f-eb0017f4c70c

+++ hale/eu.esdihumboldt.hale/src/eu/esdihumboldt/hale/rcp/views/map/style/editors/FillEditor.java
@@ -130,7 +130,7 @@ public class FillEditor implements Editor<Fill> {
 	 * @return the {@link Stroke} opacity
 	 */
 	private float getOpacity() {
-		return ((float) opacity.getSelection()) / 100.0f;
+		return (opacity.getSelection()) / 100.0f;
 	}
 
 	/**
